% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{check_rule_1}
\alias{check_rule_1}
\title{Check if more than two consecutive values in a row/column}
\usage{
check_rule_1(grid, i, j)
}
\description{
Check if more than two consecutive values in a row/column
}
