% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{check_rule_2}
\alias{check_rule_2}
\title{Check if number of 0s and 1s differ in a full row/column}
\usage{
check_rule_2(grid, i, j)
}
\description{
Check if number of 0s and 1s differ in a full row/column
}
