% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_logic.R
\name{check_rule_3}
\alias{check_rule_3}
\title{Check if any rows or columns are identical (fully filled)}
\usage{
check_rule_3(grid)
}
\description{
Check if any rows or columns are identical (fully filled)
}
